adds Okra Bridge
the okra bridge serves to wrap the okra (aka xdplayer aka pop up player)
backbone logic into an ember object. it is a place where any listeners
on backbone events can go. we can also explicitly export specific
functionality or other attributes we want to be made available to audio
service.

right now i'm not deferring sound manager intialization because it works
better. to be investigated as to why exactly.