if an Alien DOM is present, send the present document to the store
the django-page serializer will test if the object it receives is actually
a DocumentElement and set it to a new inlineDocument attribute if so.

the django-page model is updated so its document property depends on
either the original text property (as retrieved from an "api" call to
the web server) and alos the new inlineDocument prop.