Updates so django and ember can coordinate
In development, the Ember and Django apps run on separate servers, which
leads to mental overhead on our part. This update helps django
coordainte with ember so ember can point its requests for content to the
right server.

This will only happen on dev: at boot time, django will generate a file that contains
a simple node module which exports its current url. ember then consumes
that file at build/re-build time and uses the returned value in the app.

an error on the command line with a description is thrown if you don't
have the required file. a serving ember app should also reload when the
file with the django url changes.

this isn't needed on demo or prod because ember will point at its own
domain.