turns off asset fingerprinting
we'll need to turn it on to deploy the MVP in order to trigger cache
busting, but because we don't serve the ember-generated index file,
we're not going to know the asset file names at request time. I think an
apache rewrite rule can alleviate this, since the filnames always match
the pattern 'filename-<md5hash>.extension'.