updates nav-links
since these sub nav links can sometimes point to another page on the
same site, we can use a regular link-to component if the origin matches.

includes a test for the case when we're in development and the values
provided by the DB are a different origin than our configured wnycURL
value, but we know that it's the same site becaus it's development mode.