refactors poll service
we can rely on the native setInterval function if the passed function is
bound using `Ember.run.bind`, which ensures the given function is
executed within an Ember runloop. This simplifies the the scheduling
logic and allows us to maintain a single handler for every tick of the
event loop.

this commit also expands the surface area of the poll service API,
adding a number of start and stop functions, plus the option to identify
a poll with a custom label, in addition to the ID generated by the
runtime.

and this includes unit tests :)