Mirage Updates (#202)
* removes old embedded-listen-button component

* do not open external links w/target blank in testing

* use example.com instead of test.com in test config

* starts to move the calls to server out of mirage config

this was dying in dev because mirage doesn't define server in the dev
environment. the idea is to move the explicit calls to server into tests
and scenarios. we can setup a dev database with any required fixtures in
`scenarios/default` and create any required db models in related tests.

hopefully this will make thigns more explicit. there will still be some
automagical work in the django-page serializers, i.e. a call to create
a django page of id `story/foo` will make a corresponding story model
with an automatically generated id and a slug of `foo`. this basically
follows the mental model of the server, since one will never access
a djangoPage representation of a story unless that story already exists.

* adds comments to mirage and updates story

by setting up the proper serializer and models for stories and comments
in mirage, we can leverage short hand endpoints and keep the response
formatting to the serializer layer.

* create stories when a django page is generated

I have a feeling anything related to formatting the `text` attr of
a django page can be moved into a django-page serializer, and this
factory could stick to setting in-memory representations of related
story, comments, or other DB models to the generated django-page.

* explicitly generate the required django page in beta acceptance tests