adds an acceptance test for analytics
an acceptance test might be overkill, but seems like the nicest "black
box" way to test and track analytics events through the app, given the
high-level, highly integrated nature of these events.

this includes a failing test for the multiple pageview events fired when
the modal opens. need to find a way to track pageviews withouth
triggering on queryparam updates.