modernizr integration
we're using some modernizr tests in the app, but the build was removed
with all the rest of the backbone deps. this patch adds a modernizr
integration using the grunt-modernizr library, which scans our files for
uses of the modernizr API and then generates a custom build based on our
specific use-cases.

it's a little wonky because
* requires Grunt to be installed
* adds an explicit build step
* requires Grunt to be installed

would love to make this an ember add-on, but there are currently issues
about building the file and importing using the current ember-cli API.