adds Play Param mixin to support `?play` on arbitrary routes
This is a mixin we can use on any routes to support the
play-from-parameter feature on arbitrary routes.

It leverages the fact that the application controller has a query param
specific for `play` and that all routes have access to the application
controller.

Since this mixin operates on cancelable actions, we check the value
returned by the super call to see if it's explicitly false, which
indicates not to allow the event to bubble based on the route's logic.