Add ember-cli-document-title
This addon sets the titles in the routes. This means titles are ready in time for the analytics
page view event.  More reliable for our use case than ember-page-title which sets the titles with
a template helper and is not always ready in time.

(With fastboot this distinction probably wouldn't matter)