swap out `user` for `isStaff`
most uses of `user` really just needed to know if the user is a staff
member in order to show edit links. this patch swaps all of those with
the new `isStaff` value off the session service. it also passes in the
`adminURL` value from config to construct the url