unobtrusive loader is getting in the way
[ember-require-module looks up requirejs on the window](https://github.com/offirgolan/ember-require-module/blob/master/addon/index.js#L4),
which is not where unobtrusive loader defines it.

this was originally used to get around poorly behaved embedded JS that
would overwrite things like require and define on the window.

In the future if we find ourselves needing to hide the require API, we
should investigate `loader.js`'s [no conflict
feature](https://github.com/ember-cli/loader.js#no-conflict).