update show model to use relationships for image and producingOrg
this is a split from the `Channel` model, while the server reconfigures
to conform to JSON:API and use DRF to manage its endpoints.

The key for `image` is different (was `logoImage`) due to a bug in the
JSON:API extension to DRF. Included nested serialized resources at
a different key than its source throws an error. There was an issue that
was closed that was supposed to fix it, but it didn't in my experience.

https://github.com/django-json-api/django-rest-framework-json-api/pull/347