optimizes streambanner image
The streambanner image was 337K.
Persistent images from the homepage are downloaded about 189k times per
week.
The new streambanner image is 46K.
That 1/7 of the size, leading to faster download times and
approximately 55G of bandwidth savings per week.

The new image *does* reveal some JPEG quantization artifacts when
zooming in at a pixel-level, but at full size (& especially with the
grey filter on top) the differences are not immediately discernable.